requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
langchain-openai>=0.0.2.post1
pandas
python-dotenv
ipython
setuptools
streamlit 